Refactor: move :item_not_found and :merchant_not_found methods to a RecordNotFoundHelper module
