+++ app/controllers/concerns/record_not_found_helper.rb
@@ -1,0 +1,17 @@
+module RecordNotFoundHelper
+  def merchant_not_found
+    if params[:merchant_id]
+      render json: Api::V1::MerchantsController::MerchantErrorSerializer.no_merchant(params[:merchant_id]), status: 404
+    else
+      render json: Api::V1::MerchantsController::MerchantErrorSerializer.no_merchant(params[:id]), status: 404
+    end
+  end
+  
+  def item_not_found
+    if params[:item_id]
+      render json: Api::V1::ItemsController::ItemErrorSerializer.no_item(params[:item_id]), status: 404
+    else
+      render json: Api::V1::ItemsController::ItemErrorSerializer.no_item(params[:id]), status: 404
+    end
+  end
+end
